# reflect: cache IsVariadic calls in Call