--- a/src/reflect/all_test.go
+++ b/src/reflect/all_test.go
@@ -1944,17 +1944,13 @@ func BenchmarkCall(b *testing.B) {
 
 type myint int64
 
-type inccer interface {
-	inc()
-}
-
 func (i *myint) inc() {
 	*i = *i + 1
 }
 
 func BenchmarkCallMethod(b *testing.B) {
 	b.ReportAllocs()
-	z := inccer(new(myint))
+	z := new(myint)
 
 	v := ValueOf(z.inc)
 	for i := 0; i < b.N; i++ {